# swagger-doc.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "Enterprise Architecture and Project Management API with comprehensive support for frameworks, Kanban boards, Gantt charts, workflows, capability maps, ERD diagrams, and canvas visualizations.",
        "title": "ReqArchitect API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/yourusername/reqarchitect-api",
            "email": "support@reqarchitect.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/workflows": {
            "get": {
                "description": "Get all workflows for the authenticated user's organization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflow"
                ],
                "summary": "Get all workflows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "description": "Create a new workflow for an organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflow"
                ],
                "summary": "Create a new workflow",
                "parameters": [
                    {
                        "description": "Workflow object",
                        "name": "workflow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_internal_models.Workflow"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/workflows/{id}": {
            "get": {
                "description": "Get a specific workflow by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflow"
                ],
                "summary": "Get workflow by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "description": "Update an existing workflow by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflow"
                ],
                "summary": "Update a workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Updated workflow object",
                        "name": "workflow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_internal_models.Workflow"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "description": "Delete an existing workflow by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workflow"
                ],
                "summary": "Delete a workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/reqarchitect-api_pkg_utils.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "reqarchitect-api_internal_models.Framework": {
            "type": "object",
            "required": [
                "category",
                "name",
                "slug"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "config": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.JSONB"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "grandchildren": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reqarchitect-api_internal_models.FrameworkGrandchild"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "reqarchitect-api_internal_models.FrameworkGrandchild": {
            "type": "object",
            "required": [
                "frameworkId",
                "name",
                "slug"
            ],
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reqarchitect-api_internal_models.FrameworkGrandchild"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "framework": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.Framework"
                },
                "frameworkId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.JSONB"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "parent": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.FrameworkGrandchild"
                },
                "parentId": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "reqarchitect-api_internal_models.JSONB": {
            "type": "object",
            "additionalProperties": true
        },
...
        "reqarchitect-api_internal_models.Organization": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.JSONB"
                },
                "slug": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reqarchitect-api_internal_models.Team"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reqarchitect-api_internal_models.User"
                    }
                }
            }
        },
        "reqarchitect-api_internal_models.Team": {
            "type": "object",
            "required": [
                "name",
                "organizationId"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.Organization"
                },
                "organizationId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "reqarchitect-api_internal_models.User": {
            "type": "object",
            "required": [
                "email",
                "name",
                "organizationId"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.Organization"
                },
                "organizationId": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "reqarchitect-api_internal_models.Workflow": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reqarchitect-api_internal_models.WorkflowEdge"
                    }
                },
                "framework": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.Framework"
                },
                "frameworkId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reqarchitect-api_internal_models.WorkflowNode"
                    }
                },
                "organization": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.Organization"
                },
                "organizationId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "workflowType": {
                    "type": "string"
                }
            }
        },
        "reqarchitect-api_internal_models.WorkflowEdge": {
            "type": "object",
            "required": [
                "source",
                "target"
            ],
            "properties": {
                "condition": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.JSONB"
                },
                "relationshipType": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "sourceNode": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.WorkflowNode"
                },
                "target": {
                    "type": "string"
                },
                "targetNode": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.WorkflowNode"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workflow": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.Workflow"
                },
                "workflowId": {
                    "type": "string"
                }
            }
        },
        "reqarchitect-api_internal_models.WorkflowNode": {
            "type": "object",
            "required": [
                "positionX",
                "positionY",
                "type"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.JSONB"
                },
                "positionX": {
                    "type": "number"
                },
                "positionY": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workflow": {
                    "$ref": "#/definitions/reqarchitect-api_internal_models.Workflow"
                },
                "workflowId": {
                    "type": "string"
                }
            }
        },
        "reqarchitect-api_pkg_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "reqarchitect-api_pkg_utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for service-to-service authentication.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}